Fix a problem with the tpm attest certification validation
We were trying to parse 4 bytes out of the subject alternative name
manufactuerer field for the manufacturer, but the data is an 8-byte string
consisting of 8 hex-digits constituting the 4-byte code.

Additionally, it turns out we were making our `cred_id_len <= remaining_bytes`
check too late---by the time we made the check we had already consumed more
bytes out of the input.