chore: upgrade base64 to 0.21 across the workspace
this upgrades all of the crates within webauthn-rs to using base64
to the latest version 0.21. the interface changed significantly
since 0.13, and although i'm not a huge fan of it, receiving updates
is good, and this is the last crate in my workspace that needs updating
so here i am updating the crate :)

lemme know if there's anything you'd like changed with how we call
the new API's!