CredProps uses a boolean with three states (#525)
the credProps extension uses a boolean with an optional so that in
the non-present (none) cases, it represents "the inability to
determine key residency during the registration".

Chromium now decided that despite having direct hardware access to
a yubikey, where CTAP2 operations are unambiguous about the rk
state, that now my key somehow is "nebulously" in a state of both
rk and not rk at the same time.